+++ spec/lib/smartdict/commands/version_command_spec.rb
@@ -1,0 +1,5 @@
+require "spec_helper"
+
+describe Smartdict::Commands::VersionCommand do
+  it_behaves_like "command"
+end